Trying a slightly different approach to managing the media models
I'm thinking it would be better if there was just some models that encapsulate all the responses. I'm not sure if I like it this way though. I did write a test that uses a nice json package that checks if a struct has all the same fields as the json it tried to deserialize. That's pretty nice to have